requirements_min.txt
beautifulsoup4==4.9.3
bson==0.5.7
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
jsonpickle==2.0.0
Pillow==8.1.2
powernad==0.8.0
pymongo==3.7.2
requests==2.25.1